wip... pushing progress for mobile migration to typescript
