+++ FA22-P05-Mobile/App.js
@@ -1,20 +1,19 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
-
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>Open up App.js to start working on your app!</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+"use strict";
+exports.__esModule = true;
+var expo_status_bar_1 = require("expo-status-bar");
+var react_native_1 = require("react-native");
+function App() {
+    return (<react_native_1.View style={styles.container}>
+      <react_native_1.Text>Open up App.js to start working on your app!</react_native_1.Text>
+      <expo_status_bar_1.StatusBar style="auto"/>
+    </react_native_1.View>);
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
+exports["default"] = App;
+var styles = react_native_1.StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center'
+    }
 });
