+++ FA22-P05-Mobile/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    preset: 'react-native',
+    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node']
+  };
