+++ FA22-P05-Mobile/package.json
@@ -15,7 +15,14 @@
     "react-native": "0.69.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9"
+    "@babel/core": "^7.12.9",
+    "@types/jest": "^29.1.1",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-native": "^0.70.4",
+    "@types/react-test-renderer": "^18.0.0",
+    "tslint": "^6.1.3",
+    "typescript": "^4.8.4"
   },
   "private": true
 }
